feat: add prettier and eslint
