+++ .env
@@ -1,0 +1,3 @@
+PORT=3000
+TOKEN='Bearer 123456'
+REFRESH_TOKEN='refresh token'
