+++ .prettierrc
@@ -1,0 +1,7 @@
+
+{
+    "semi": true,
+    "singleQuote": true,
+    "tabWidth": 2,
+    "useTabs": false
+}
